refactor todo list with uuid
