+++ package.json
@@ -16,6 +16,7 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-countdown-circle-timer": "^3.0.9",
+    "react-native-get-random-values": "^1.7.2",
     "react-native-web": "0.17.1",
     "uuid": "^8.3.2",
     "uuidv4": "^6.2.12"
